feature: Display Shell Prompt and Read Input
This commit adds support for displaying the shell prompt and
reading/storing input from the command line. The shell uses the PS1
variable as the prompt string if it is set as an environment variable
before starting the shell. If it is not, then the shell uses its own
customizable prompt string.

The custom prompt string is written to stdout by ExpandPromptString().
This function prints the custom prompt string, kPromptString, and
replaces any espace codes included. As of now, this implementation only
supports the following escape code:

  - \u: System username that started the shell
  - \h: System hostname
  - \b: Basename of the current working directory