Commit Dynamic Array Implementation
This commit introduces a basic implementation of a dynamic array. For
improved flexibility, this data structure is type agnostic. To account
for any type, the size in bytes of the type must be specified when
initializing the array in order to properly manage memory allocations
and access the data.

This data structure was implemented to simplify storing the tokenized
command line input and provide greater flexibility when parsing the
input.