feature: Implement cd and exit
Given that changing the shell's current directory and exiting must
change the state of the shell, these functions were built into it.
Otherwise, executing these commands from a forked process would change
the state of the child process but not that of the shell.

This implementation assumes that builtin commands are the first argument
of the command line. Otherwise, they are treated as arguments of another
command or pathnames.