feature: Implement Process type
The `Process` structure contains information related to a process
executed by the shell. This includes values that are passed to exec,
such as the path to the command to be executed and a `char *` array with
its arguments. In case of redirection, the process stores redirected
file descriptors and a copy of the original standard streams file
descriptors (used to undo any redirection).