feature: Commit Redirection Functions
To simplify the process of identifying and handling redirection
operators, an enum type RedirectType was defined. In conjunction with
`GetRedirectType()`, which maps string character with RedirectType,
these two are used to identify whether an argument in the command line
corresponds to a redirection operator.

On the other hand, `SetupRedirection()` is used to redirect a given
standard stream to a new file descriptor. The duplicated file
descriptors are stored in the corresponding std* attribute of a Process
structure.

The `CleanupRedirection()` function is used to efficiently and securely
restore any redirection and close duplicated file descriptors.

When `SetupRedirection()` encounters a failure, it reverses any
successful stream redirections by restoring the original streams and
closing any duplicated file descriptors. This approach ensures
consistency, as incomplete redirections would render a command unable
to fulfill its expected function.