bugfix: Fix AppendElement() Resize Condition and Error Checking
Array was being resized when its size was greater than the number of
elements in the array, which is always true. Additionally, error
checking around `ResizeDynamicArray()` was incorrect, since it was
written under the misconception that this function returned a pointer.
However, it returns an integer instead. As a result, a successful return
value (0) was interpreted as an error by `AppendElement()`.