bugfix: Close Redirected File Descriptors on Multiple Redirections
In the case of multiple redirections, the program would override
previous file descriptors without closing them first. This leads to an
unefficient management of resources. To address them, before setting up
a redirection, checks were added to ensure that if there is a file
descriptor that was already used for redirection and will be
overwritten, the old file descriptor is closed before being replaced.