feature: Implement Command Line Parsing Logic
This commit implements `ParseCommand()`, whose purpose is to parse the
tokenized arguments from the command line and perform two operations:

  1. Setup redirection: Identifies redirection operators and its target
     file, and applies the redirection with the use of redirection
     functions.
  2. Extract command and arguments: Every time a redirection is
     setup, the function removes the redirection operator and the target
     file. Once the tokenized arguments have been parsed, the command
     and its arguments are stored in the Process structure for later
     execution.

Given the complexity of the function when it comes to handling errors,
delegating error reporting to the caller causes the error messages to be
more general. This is because the error reported by the caller would be
generated with `errno`, missing important information that might
illustrate the cause of an error during parsing. For this reason, errors
are reported by `ParseCommand()` directly to maintain the specificity
of error messages and avoid implementing other unecessarily complex
error reporting methods.