bugfix: Fix Open Flags for Append Redirection
Previously, the flags specified to `open` for handling append
redirection were `O_CREAT`, `O_TRUNC`, and `O_APPEND`. Using the later
two flags seems counterintuitive, as we are asking the system to empty
the file and append to it. The combination of these two flags caused
writing to a file with `>>` redirection operator to raise a 'Bad file
descriptor' error. This indicates an attempt to write to a file
descriptor that is not valid for writing.

Opening the file with `O_CREAT`, `O_APPEND`, `O_WRONLY` when the append
redirection operator is used solved the issue.