feature: Support Shell Variable '$?'
This commit introduces the capability to check the exit status of the
last executed command. When using `echo` to display this shell variable,
the program replaces `$?` during the parsing process with the global
variable `status` using the function `ReplaceExitStatusVariable()`.

Note: `$?` is interpreted as a shell variable when used with `echo`.
Otherwise, it is treated as a string literal. Since this current
implementation does not support Bash scripting, the only reasonable case
to use `$?` is with `echo`.